I changed several things in this draft, but it is still broken in some ways.
In particular, I changed some indices on several maps, including the construction of kappa. I also changed the way that kappaLifted is constructed with the hopes of solving the issue at the end that the source and target maps of f_i don't line up with the modules they're supposed to be.

Unfortunately, this hasn't actually solved the problem since cRes still says that there are the wrong number of rows and/or columns. I'll have to look at this more later.